Trait ispc_rt::instrument::Instrument

pub trait Instrument {
    fn instrument(
        &self, 
        file: &CStr, 
        note: &CStr, 
        line: i32, 
        mask: u64, 
        active_count: u32
    );

    fn print_summary(&self) { ... }
}

Trait to be implemented to provide ISPC instrumentation functionality.

The runtime required function is wrapped and forwarded to your struct.

Required methods

fn instrument(
&self,
file: &CStr,
note: &CStr,
line: i32,
mask: u64,
active_count: u32
)

instrument is called when ISPC calls the ISPCInstrument callback. The file and note strings are converted CStr and the number of active programs is computed from the mask.

Loading content...
Provided methods

fn print_summary(&self)

Called through ispc::print_instrumenting_summary, optionally log out a summary of performance information gathered through the instrument callback.

Loading content...