Struct ispc_rt::task::Context

pub struct Context {
    pub id: usize,
    // some fields omitted
}

A list of all task groups spawned by a function in some launch context which will be sync'd at an explicit sync call or function exit.

Note: A Context is done if and only if ISPCSync has been called with its handle and all of its tasks are finished. Until ISPCSync is called on the Context's handle more tasks could be launched.

Additionally, because we're not really able to associate a call to ISPCAlloc with a specific Group care must be taken that the Context is not dropped until ISPCSync has been called on its handle and all Groups within have completed execution.

Fields
id: usize

A unique identifier for this context

Methods

impl Context

pub fn new(id: usize) -> Context

Create a new list of tasks for some function with id id

pub fn launch(&self, total: (i32, i32, i32), data: *mut c_void, fcn: ISPCTaskFn)

Add a task group for execution that was launched in this context

pub fn current_tasks_done(&self) -> bool

Check if all tasks currently in the task list are completed

Note: A Context is done if and only if ISPCSync has been called with its handle and all of its tasks are finished. Until ISPCSync is called on the Context's handle more tasks could be launched. TODO: With this design we're essentially requiring the thread waiting on the context to busy wait since we provide no condition variable to block on.

pub unsafe fn alloc(&self, size: usize, align: usize) -> *mut c_void

Allocate some memory for this Context's task groups, returns a pointer to the allocated memory.

Important traits for ContextIter<'a>
impl<'a> Iterator for ContextIter<'a> type Item = Arc<Group>;

pub fn iter(&self) -> ContextIter

An iterator over the current groups in the context which have remaining tasks to run on a thread. If more task groups are added before this iterator has returned None those will appear as well.